pip_requirements.txt
Mastodon.py==1.4.3
beautifulsoup4==4.7.1

